+++ src/main/kotlin/org/jetbrains/intellij/IntelliJPlugin.kt
@@ -634,6 +634,7 @@ open class IntelliJPlugin : Plugin<Project> {
 
             jbrResolver.resolveRuntime(
                 jbrVersion = task.jbrVersion.orNull,
+                jbrVariant = task.jbrVariant.orNull,
                 ideDir = task.ideDir.orNull,
             )
         })